refactor(config): consolidate config management and improve error handling
