+++ src/libs/app-error.ts
@@ -16,6 +16,7 @@ export const AppErrorCode = {
 	HOOK_FAILED: "HOOK_FAILED",
 	PATH_INVALID: "PATH_INVALID",
 	CANCELLED: "CANCELLED",
+	INVALID_INPUT: "INVALID_INPUT",
 	INTERNAL: "INTERNAL",
 } as const;
