+++ src/libs/command-error-handler.ts
@@ -1,12 +1,12 @@
 import { Result } from "typescript-result";
-import { presentCommandError } from "../cmds/present-error.ts";
+import { presentCommandError, type PresentErrorOptions } from "../cmds/present-error.ts";
 
 export interface ErrorHandler {
-	onError(error: Error, commandName: string, options?: { debug?: boolean }): void;
+	onError(error: Error, commandName: string, options?: PresentErrorOptions): void;
 }
 
 export class ConsoleErrorHandler implements ErrorHandler {
-	onError(error: Error, commandName: string, options?: { debug?: boolean }): void {
+	onError(error: Error, commandName: string, options?: PresentErrorOptions): void {
 		presentCommandError(commandName, error, options);
 	}
 }
@@ -14,7 +14,7 @@ export class ConsoleErrorHandler implements ErrorHandler {
 export class CommandErrorHandler {
 	constructor(private readonly errorHandler: ErrorHandler) {}
 
-	handle<T>(result: Result<T, Error>, commandName: string, options?: { debug?: boolean }): T | null {
+	handle<T>(result: Result<T, Error>, commandName: string, options?: PresentErrorOptions): T | null {
 		if (!result.ok) {
 			this.errorHandler.onError(result.error, commandName, options);
 			return null;
@@ -25,14 +25,14 @@ export class CommandErrorHandler {
 	async handleAsync<T>(
 		promise: Promise<Result<T, Error>>,
 		commandName: string,
-		options?: { debug?: boolean },
+		options?: PresentErrorOptions,
 	): Promise<T | null> {
 		const result = await promise;
 		return this.handle(result, commandName, options);
 	}
 
 	// Static factory methods for convenience
-	static withExit<T>(result: Result<T, Error>, commandName: string, options?: { debug?: boolean }): T | null {
+	static withExit<T>(result: Result<T, Error>, commandName: string, options?: PresentErrorOptions): T | null {
 		if (!result.ok) {
 			presentCommandError(commandName, result.error, options);
 			Deno.exit(1);
@@ -43,7 +43,7 @@ export class CommandErrorHandler {
 	static async withExitAsync<T>(
 		promise: Promise<Result<T, Error>>,
 		commandName: string,
-		options?: { debug?: boolean },
+		options?: PresentErrorOptions,
 	): Promise<T | null> {
 		const result = await promise;
 		return this.withExit(result, commandName, options);
